Arreglados bug y code smells
