+++ jest.config.cjs
@@ -1,4 +1,11 @@
 module.exports = {
   testEnvironment: 'jest-environment-jsdom',
-  setupFiles: ['./jest.setup.js']
+  setupFiles: ['./jest.setup.js'],
+  collectCoverageFrom: [
+    'src/**/*',
+    '!src/*',
+    '!src/app/**/*',
+    '!src/assets/**/*',
+    '!src/routes/**/*'
+  ]
 }
