+++ sonar-project.properties
@@ -7,7 +7,7 @@ sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 sonar.inclusions=src/**/*
 
 # Excluded files
-sonar.exclusions=src/app/**/*,src/assets/**/*,src/routes/**/*,src/*.{js,jsx}
+sonar.exclusions=src/*,src/app/**/*,src/assets/**/*,src/routes/**/*
 
 # This is the name and version displayed in the SonarCloud UI.
 #sonar.projectName=frabricafrontend
