+++ src/views/login/Login.css
@@ -1,7 +1,6 @@
 body {
     margin: 0;
     padding: 0;
-    font-family: montserrat;
     background: linear-gradient(120deg, #43ba5e, #3fb09b);
     height: 100vh;
     overflow: hidden;
